+++ R/AddCardModule.R
@@ -45,13 +45,15 @@ add_card_button_ui <- function(id) {
 #' @param id `character(1)` this `shiny` module's id.
 #' @param reporter [`Reporter`] instance.
 #' @param card_fun `function` which returns a [`ReportCard`] instance,
-#' the function have at`card`argument and optional `comment`.
+#' the function has optional `card` and  `comment` arguments.
+#' If the `card` argument is added then the `ReportCard` instance is automatically created for the user.
+#' If the comment argument is not specified then it is added automatically at the end of the Card.
 #' @return `shiny::moduleServer`
 #' @export
 add_card_button_srv <- function(id, reporter, card_fun) {
   checkmate::assert_function(card_fun)
   checkmate::assert_class(reporter, "Reporter")
-  checkmate::assert_subset(names(formals(card_fun)), c("card", "comment"), empty.ok = FALSE)
+  checkmate::assert_subset(names(formals(card_fun)), c("card", "comment"), empty.ok = TRUE)
 
   shiny::moduleServer(
     id,
@@ -101,25 +103,29 @@ add_card_button_srv <- function(id, reporter, card_fun) {
 
       shiny::observeEvent(input$add_card_ok, {
         card_fun_args_nams <- names(formals(card_fun))
-        # The default_card is defined here because formals() returns a pairedlist object
-        # of formal parameter names and their default values. The values are missing
-        # if not defined and the missing check does not work if supplied formals(card_fun)[[1]]
-        default_card <- formals(card_fun)[[1]]
-        card <- `if`(
-          missing(default_card),
-          ReportCard$new(),
-          eval(default_card, envir = environment(card_fun))
-        )
-        has_own_comment <- length(card_fun_args_nams) == 2
-        if (has_own_comment) {
-          card <- try(card_fun(card, input$comment))
+        has_card_arg <- "card" %in% card_fun_args_nams
+        has_comment_arg <- "comment" %in% card_fun_args_nams
+
+        if (!has_card_arg) {
+          if (has_comment_arg) {
+            card <- try(card_fun(comment = input$comment))
+          } else {
+            card <- try(card_fun())
+          }
         } else {
-          card <- try(card_fun(card))
-          if (!inherits(card, "try-error")) {
-            if (length(input$comment) > 0 && input$comment != "") {
-              card$append_text("Comment", "header3")
-              card$append_text(input$comment)
-            }
+          # The default_card is defined here because formals() returns a pairedlist object
+          # of formal parameter names and their default values. The values are missing
+          # if not defined and the missing check does not work if supplied formals(card_fun)[[1]]
+          default_card <- formals(card_fun)[[1]]
+          card <- `if`(
+            missing(default_card),
+            ReportCard$new(),
+            eval(default_card, envir = environment(card_fun))
+          )
+          if (has_comment_arg) {
+            card <- try(card_fun(card = card, comment = input$comment))
+          } else {
+            card <- try(card_fun(card = card))
           }
         }
 
@@ -136,6 +142,10 @@ add_card_button_srv <- function(id, reporter, card_fun) {
           )
         } else {
           checkmate::assert_class(card, "ReportCard")
+          if (!has_comment_arg && length(input$comment) > 0 && input$comment != "") {
+            card$append_text("Comment", "header3")
+            card$append_text(input$comment)
+          }
           reporter$append_cards(list(card))
           shiny::showNotification(sprintf("The card added successfully."), type = "message")
           shiny::removeModal()