+++ NEWS.md
@@ -1,7 +1,11 @@
 # teal.reporter 0.1.0.9005
 
+* Added the `Archiver` class which enables saving and loading back the `Report` in the `shiny` session.
+* Added `to_list` and `from_list` methods to all content related classes.
 * Added support for the `ElementaryTree` class in the `append_table` method of `ReportCard`.
 * Added additional validation for the `card_fun` evaluation.
+* Decoupled the `metadata` and `content` in the `ReportCard`.
+* Added support for more arguments setup for a `card_fun` function in the `add_card_button_srv` module.
 
 # teal.reporter 0.1.0